+++ src/main/java/dk/mada/jaxrs/generator/dto/DtoGenerator.java
@@ -29,14 +29,10 @@ import dk.mada.jaxrs.generator.dto.tmpl.CtxEnum;
 import dk.mada.jaxrs.generator.dto.tmpl.CtxExtraDateSerializer;
 import dk.mada.jaxrs.generator.dto.tmpl.CtxInterface;
 import dk.mada.jaxrs.generator.dto.tmpl.CtxProperty;
-import dk.mada.jaxrs.generator.dto.tmpl.CtxPropertyExt;
 import dk.mada.jaxrs.generator.imports.Imports;
 import dk.mada.jaxrs.generator.imports.Jackson;
 import dk.mada.jaxrs.generator.imports.JavaIo;
-import dk.mada.jaxrs.generator.imports.JavaMath;
-import dk.mada.jaxrs.generator.imports.JavaUtil;
 import dk.mada.jaxrs.generator.imports.UserMappedImport;
-import dk.mada.jaxrs.generator.imports.ValidationApi;
 import dk.mada.jaxrs.model.Dto;
 import dk.mada.jaxrs.model.Dtos;
 import dk.mada.jaxrs.model.Info;
@@ -45,16 +41,8 @@ import dk.mada.jaxrs.model.Property;
 import dk.mada.jaxrs.model.SubtypeSelector;
 import dk.mada.jaxrs.model.types.Primitive;
 import dk.mada.jaxrs.model.types.Type;
-import dk.mada.jaxrs.model.types.TypeArray;
-import dk.mada.jaxrs.model.types.TypeByteArray;
-import dk.mada.jaxrs.model.types.TypeContainer;
-import dk.mada.jaxrs.model.types.TypeEnum;
 import dk.mada.jaxrs.model.types.TypeInterface;
-import dk.mada.jaxrs.model.types.TypeMap;
-import dk.mada.jaxrs.model.types.TypeReference;
-import dk.mada.jaxrs.model.types.TypeSet;
 import dk.mada.jaxrs.naming.Naming;
-import dk.mada.jaxrs.openapi.OpenapiGeneratorUtils;
 
 /**
  * DTO generator.
@@ -64,8 +52,6 @@ import dk.mada.jaxrs.openapi.OpenapiGeneratorUtils;
 public class DtoGenerator {
     private static final Logger logger = LoggerFactory.getLogger(DtoGenerator.class);
 
-    /** Naming. */
-    private final Naming naming;
     /** Types. */
     private final Dtos dtos;
     /** Generator options. */
@@ -87,6 +73,8 @@ public class DtoGenerator {
 
     /** Enumeration generator. */
     private final EnumGenerator enumGenerator;
+    /** Property generator. */
+    private PropertyGenerator propertyGenerator;
 
     /**
      * Constructs a new generator.
@@ -97,7 +85,6 @@ public class DtoGenerator {
      * @param model     the data model
      */
     public DtoGenerator(Naming naming, GeneratorOpts opts, Templates templates, Model model) {
-        this.naming = naming;
         this.opts = opts;
         this.templates = templates;
         this.model = model;
@@ -106,6 +93,7 @@ public class DtoGenerator {
         externalTypeMapping = opts.getExternalTypeMapping();
 
         enumGenerator = new EnumGenerator(naming, opts);
+        propertyGenerator = new PropertyGenerator(naming, opts, enumGenerator);
     }
 
     /**
@@ -225,7 +213,7 @@ public class DtoGenerator {
         CtxEnum ctxEnum = null;
         if (isEnum) {
             List<String> enumValues = dto.enumValues();
-            ctxEnum = enumGenerator.buildEnumEntries(dtoType, enumValues);
+            ctxEnum = enumGenerator.toCtxEnum(dtoType, enumValues);
             enumSchema = enumGenerator.buildEnumSchema(dtoImports, dtoType, ctxEnum);
         }
 
@@ -316,7 +304,7 @@ public class DtoGenerator {
         Comparator<? super CtxProperty> propertySorter = propertySorter();
 
         Stream<CtxProperty> props = ds.properties().stream()
-                .map(p -> toCtxProperty(ds.imports(), p));
+                .map(p -> propertyGenerator.toCtxProperty(ds.imports(), p));
 
         if (propertySorter != null) {
             props = props.sorted(propertySorter);
@@ -445,285 +433,7 @@ public class DtoGenerator {
         return Optional.of(implementsInterfaces);
     }
 
-
-    private CtxProperty toCtxProperty(Imports dtoImports, Property p) {
-        final String name = p.name();
-        final String varName = naming.convertPropertyName(name);
-
-        String nameCamelized = OpenapiGeneratorUtils.camelize(varName);
-        String nameSnaked = OpenapiGeneratorUtils.underscore(nameCamelized).toUpperCase();
-        // Both Jackson (Fasterxml) and JsonBinding expect the getter
-        // of a 'xX'-prefixed field to be named 'getxX'. Although
-        // this is different from Bean Spec naming for getters/setters.
-        //
-        // See https://github.com/FasterXML/jackson-databind/blob/2.15...
-        // /src/main/java/com/fasterxml/jackson/databind/introspect...
-        // /DefaultAccessorNamingStrategy.java#L182
-        if (name.length() > 1 && Character.isUpperCase(name.charAt(1))) {
-            nameCamelized = Character.toLowerCase(name.charAt(0)) + nameCamelized.substring(1);
-        }
-
-        logger.debug("Property {} -> {} / {} / {}", name, varName, nameCamelized, nameSnaked);
-
-        Type propType = p.reference().refType();
-        logger.trace(" {}", propType);
-
-        String defaultValue = null;
-        boolean isRequired = p.isRequired();
-        boolean isByteArray = false;
-        boolean isArray = false;
-        boolean isMap = false;
-        boolean isSet = false;
-        boolean isDate = propType.isDate();
-        boolean isDateTime = propType.isDateTime();
-        boolean isEnum = false;
-        String innerTypeName = null;
-        CtxEnum ctxEnum = null;
-        Type innerType = null;
-
-        if (propType instanceof TypeByteArray) {
-            isByteArray = true;
-            dtoImports.add(JavaUtil.ARRAYS);
-            if (isRequired) {
-                defaultValue = "new byte[] {}";
-            }
-        }
-        if (propType instanceof TypeArray ca) {
-            isArray = true;
-            innerType = ca.innerType();
-            defaultValue = "new " + ca.containerImplementation() + "<>()";
-        }
-        if (propType instanceof TypeMap cm) {
-            isMap = true;
-            innerType = cm.innerType();
-            defaultValue = "new " + cm.containerImplementation() + "<>()";
-        }
-        if (propType instanceof TypeSet cs) {
-            isSet = true;
-            innerType = cs.innerType();
-            defaultValue = "new " + cs.containerImplementation() + "<>()";
-
-            // In templates, array is used for both set and list
-            isArray = true;
-        }
-        if (propType.isEnum()) {
-            innerType = propType;
-        }
-
-        if (innerType != null) {
-            innerTypeName = innerType.typeName().name();
-        }
-
-        final String typeName = propType.wrapperTypeName().name();
-        final boolean isContainer = isArray || isMap || isSet;
-        String enumClassName = typeName;
-        String enumTypeName = typeName;
-        Optional<String> enumSchema = Optional.empty();
-
-        if (getDereferencedInnerEnumType(innerType)instanceof TypeEnum te) {
-            isEnum = true;
-            Type enumType = te.innerType();
-            enumTypeName = enumType.typeName().name();
-            enumClassName = te.typeName().name();
-            dtoImports.addEnumImports(!isContainer, !isTypePrimitiveEquals(enumType));
-
-            ctxEnum = enumGenerator.buildEnumEntries(enumType, te.values());
-            enumSchema = enumGenerator.buildEnumSchema(dtoImports, enumType, ctxEnum);
-
-            logger.debug(" enum {} : {}", innerTypeName, te.values());
-        }
-
-        // Add import if required
-        UserMappedImport externalType = externalTypeMapping.get(typeName);
-        if (externalType != null) {
-            dtoImports.add(externalType);
-        }
-        if (innerTypeName != null) {
-            UserMappedImport innerExternalType = externalTypeMapping.get(innerTypeName);
-            if (innerExternalType != null) {
-                dtoImports.add(innerExternalType);
-            }
-        }
-
-        String getterPrefix = getterPrefix(p);
-        String getter = getterPrefix + nameCamelized;
-        String setter = "set" + nameCamelized;
-        String extGetter = getter;
-        String extSetter = setter;
-        boolean isUseBigDecimalForDouble = opts.isUseBigDecimalForDouble()
-                && propType.isPrimitive(Primitive.DOUBLE);
-        if (isUseBigDecimalForDouble) {
-            getter = getter + "Double";
-            setter = setter + "Double";
-
-            dtoImports.add(Jackson.JSON_IGNORE);
-            dtoImports.add(JavaMath.BIG_DECIMAL);
-        }
-
-        Optional<String> description = p.description();
-
-        boolean isUseEmptyCollections = opts.isUseEmptyCollections()
-                && isContainer
-                && !isRequired;
-        if (isUseEmptyCollections) {
-            dtoImports.add(Jackson.JSON_IGNORE);
-            getter = getter + "Nullable";
-        }
-
-        List<String> schemaEntries = new ArrayList<>();
-        if (isRequired) {
-            schemaEntries.add("required = true");
-        }
-        if (p.isNullable()) {
-            schemaEntries.add("nullable = true");
-        }
-        if (p.isReadonly()) {
-            schemaEntries.add("readOnly = true");
-        }
-        consumeNonBlankEncoded(description, d -> schemaEntries.add("description = \"" + d + "\""));
-        consumeNonBlankEncoded(p.example(), e -> schemaEntries.add("example = \"" + e + "\""));
-
-        Optional<String> schemaOptions = Optional.empty();
-        if (!schemaEntries.isEmpty()) {
-            schemaOptions = Optional.of(String.join(", ", schemaEntries));
-            dtoImports.addMicroProfileSchema();
-        }
-
-        boolean useBeanValidation = opts.isUseBeanValidation();
-        boolean valid = false;
-        Optional<String> minLength = Optional.empty();
-        Optional<String> maxLength = Optional.empty();
-        Optional<String> minimum = Optional.empty();
-        Optional<String> maximum = Optional.empty();
-        Optional<String> decimalMinimum = Optional.empty();
-        Optional<String> decimalMaximum = Optional.empty();
-        Optional<String> pattern = Optional.empty();
-        if (useBeanValidation) {
-            if (p.isRequired()) {
-                dtoImports.add(ValidationApi.NOT_NULL);
-            }
-            // Decide where to put @Valid. I expect this to be too simple...
-            if (propType.isDto()
-                    || (propType instanceof TypeContainer tc && tc.innerType().isDto())) {
-                valid = true;
-                dtoImports.add(ValidationApi.VALID);
-            }
-
-            // Note that OpenApi specification xItems/xLength both map to @Size
-            minLength = p.minItems()
-                    .or(p::minLength)
-                    .map(i -> Integer.toString(i)); // NOSONAR - not enough information to select variant
-            if (minLength.isPresent()) {
-                dtoImports.add(ValidationApi.SIZE);
-            }
-            maxLength = p.maxItems()
-                    .or(p::maxLength)
-                    .map(i -> Integer.toString(i)); // NOSONAR - not enough information to select variant
-            if (maxLength.isPresent()) {
-                dtoImports.add(ValidationApi.SIZE);
-            }
-
-            if (propType.isBigDecimal()) {
-                decimalMinimum = p.minimum()
-                        .map(min -> "\"" + min.toString() + "\"");
-                decimalMaximum = p.maximum()
-                        .map(max -> "\"" + max.toString() + "\"");
-            } else {
-                minimum = p.minimum()
-                        .map(min -> Long.toString(min.longValue()));
-                maximum = p.maximum()
-                        .map(max -> Long.toString(max.longValue()));
-            }
-            if (decimalMinimum.isPresent()) {
-                dtoImports.add(ValidationApi.DECIMAL_MIN);
-            }
-            if (minimum.isPresent()) {
-                dtoImports.add(ValidationApi.MIN);
-            }
-            if (decimalMinimum.isPresent()) {
-                dtoImports.add(ValidationApi.DECIMAL_MAX);
-            }
-            if (maximum.isPresent()) {
-                dtoImports.add(ValidationApi.MAX);
-            }
-
-            pattern = p.pattern()
-                    .map(StringRenderer::encodeRegexp);
-            if (pattern.isPresent()) {
-                dtoImports.add(ValidationApi.PATTERN);
-            }
-        }
-
-        CtxPropertyExt mada = CtxPropertyExt.builder()
-                .innerDatatypeWithEnum(innerTypeName)
-                .enumClassName(enumClassName)
-                .enumTypeName(enumTypeName)
-                .enumSchemaOptions(enumSchema)
-                .schemaOptions(schemaOptions)
-                .isUseBigDecimalForDouble(isUseBigDecimalForDouble)
-                .isUseEmptyCollections(isUseEmptyCollections)
-                .isByteArray(isByteArray)
-                .isEqualsPrimitive(isTypePrimitiveEquals(propType))
-                .getter(extGetter)
-                .setter(extSetter)
-                .jsonb(opts.isJsonb())
-                .valid(valid)
-                .renderJavadocMacroSpacer(description.isPresent())
-                .build();
-
-        CtxProperty prop = CtxProperty.builder()
-                .baseName(name)
-                .datatypeWithEnum(typeName)
-                .dataType(innerTypeName)
-                .name(varName)
-                .nameInCamelCase(nameCamelized)
-                .nameInSnakeCase(nameSnaked)
-                .getter(getter)
-                .setter(setter)
-                .description(description.flatMap(StringRenderer::makeValidPropertyJavadocSummary))
-                .isArray(isArray)
-                .isEnum(isEnum)
-                .isMap(isMap)
-                .isSet(isSet)
-                .isContainer(isContainer)
-                .isDate(isDate)
-                .isDateTime(isDateTime)
-                .defaultValue(defaultValue)
-                .required(isRequired)
-                .example(p.example())
-                .allowableValues(ctxEnum)
-                .useBeanValidation(useBeanValidation)
-                .minLength(minLength)
-                .maxLength(maxLength)
-                .minimum(minimum)
-                .maximum(maximum)
-                .decimalMinimum(decimalMinimum)
-                .decimalMaximum(decimalMaximum)
-                .pattern(pattern)
-                .madaProp(mada)
-                .build();
-
-        logger.debug("property {} : {}", name, prop);
-        return prop;
-    }
-
     private boolean isTypePrimitiveEquals(Type t) {
         return t.isPrimitive(Primitive.INT);
-    }
-
-    private @Nullable Type getDereferencedInnerEnumType(@Nullable Type t) {
-        if (t instanceof TypeReference tr) {
-            return tr.refType();
-        }
-        return t;
-    }
-
-    private String getterPrefix(Property p) {
-        boolean isBoolean = p.reference().isPrimitive(Primitive.BOOLEAN);
-        String getterPrefix = "get";
-        if (isBoolean && !opts.isUseBooleanGetPrefix()) {
-            getterPrefix = "is";
-        }
-        return getterPrefix;
     }
 }
